+++ libs/components/src/lib/components/overlay/components/toggletip/stories/components/index.ts
@@ -1,0 +1,1 @@
+export * from './toggletip-storybook.component';