+++ libs/components/src/lib/components/sort/partials/sort/public-api.ts
@@ -1,0 +1,2 @@
+export * from './sort.directive';
+export * from './sort.types';