+++ libs/components/src/lib/components/table/partials/cells/cell/cell.directive.ts
@@ -1,0 +1,11 @@
+import { CdkCell } from '@angular/cdk/table';
+import { Directive } from '@angular/core';
+
+@Directive({
+  selector: 'et-cell, td[et-cell]',
+  host: {
+    class: 'et-cell et-data-table__cell',
+  },
+  standalone: true,
+})
+export class CellDirective extends CdkCell {}