+++ libs/components/src/lib/components/table/partials/cells/header-cell-def/public-api.ts
@@ -1,0 +1,1 @@
+export * from './header-cell-def.directive';