+++ libs/contentful/src/lib/components/rich-text-renderer/public-api.ts
@@ -1,0 +1,3 @@
+export * from './rich-text-renderer.component';
+export * from './rich-text-renderer.util';
+export * from './rich-text-renderer.types';