+++ libs/contentful/src/lib/components/rich-text-renderer/rich-text-renderer.types.ts
@@ -1,0 +1,6 @@
+export interface RichTextRenderCommand {
+  payload: string | (new () => unknown);
+  data?: unknown;
+  attributes?: Record<string, string>;
+  children: RichTextRenderCommand[];
+}