+++ libs/contentful/src/lib/constants/contentful.constants.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { ContentfulConfig } from '../utils';
+
+export const CONTENTFUL_CONFIG = new InjectionToken<ContentfulConfig>('TooltipConfig');