+++ libs/core/src/lib/directives/public-api.ts
@@ -1,0 +1,5 @@
+export * from './let/public-api';
+export * from './repeat/public-api';
+export * from './observe-resize/public-api';
+export * from './observe-content/public-api';
+export * from './click-outside/public-api';