+++ libs/core/src/lib/public-api.ts
@@ -1,4 +1,7 @@
 export * from './constants/public-api';
+export * from './decorators/public-api';
+export * from './directives/public-api';
+export * from './pipes/public-api';
 export * from './services/public-api';
 export * from './types/public-api';
 export * from './utils/public-api';