+++ libs/components/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @ethlete/components
 
+## 0.2.0-next.1
+
+### Patch Changes
+
+- [#65](https://github.com/ethlete-io/ethdk/pull/65) [`1cf9a35`](https://github.com/ethlete-io/ethdk/commit/1cf9a35a39885c1054721db20ce23424c81bad74) Thanks [@TomTomB](https://github.com/TomTomB)! - Add missing exports
+
+## 0.2.0-next.0
+
+### Minor Changes
+
+- [#49](https://github.com/ethlete-io/ethdk/pull/49) [`5c95058`](https://github.com/ethlete-io/ethdk/commit/5c9505837ee3e5f2457169591acd01c79eade565) Thanks [@TomTomB](https://github.com/TomTomB)! - - Add table component
+
+  - Add pagination component
+  - Add skeleton component
+  - Add sort component
+  - Add tooltip component
+
+- [#35](https://github.com/ethlete-io/ethdk/pull/35) [`c94046b`](https://github.com/ethlete-io/ethdk/commit/c94046b64ff2986f8518bed7cc9a4f5104bd7972) Thanks [@TomTomB](https://github.com/TomTomB)! - Add dynamic overlay service
+
 ## 0.1.0
 
 ### Minor Changes